Introduce `audio_dac` component
