--- esphome/components/audio_processor/__init__.py
@@ -1,12 +1,0 @@
-import esphome.codegen as cg
-from esphome.core import coroutine_with_priority
-
-CODEOWNERS = ["@kbx81"]
-IS_PLATFORM_COMPONENT = True
-
-audio_processor_ns = cg.esphome_ns.namespace("audio_processor")
-
-
-@coroutine_with_priority(100.0)
-async def to_code(config):
-    pass
